Responsive menu logic to change menu behaviors based on screen width.
The changes:
An event listener on window listens for changes to the screen size. At the prescribed width of 860px, the user can no longer close dropdown menus by pressing on an element that is not it.

Meanwhile, for #toggleDropdown(), if the mobile breakpoint condition is met, then we allow the user to have multiple dropdown menus open at once. Otherwise, only one at a time.