Slider CSS
Key notes:
I had to set the width and the height for the slider manually; thereafter I could use % values for width and height for tis children. This is extremely important for keeping the image and the frame within the .slider container.

Easily position the caption using absolute and left/rigth centering with bottom px.